// scripts/vendor/siwc-local/types.ts
export interface ChatGPTModel {
  slug: string;
  displayName: string;
}